replace MinioService with multiple service interfaces
- Removed the monolithic MinioService interface and implementation.
- Introduced separate service interfaces for FileService, FolderService,
 SearchService, and StorageUsageService.
- Created abstract base class AbstractMinioService for common
functionality.
- Implemented specific service classes (e.g., MinioFileServiceImpl) for
each service interface.
- Updated controller dependencies to use new service interfaces.
- Moved utility methods to PathUtil for better separation of concerns.
