+++ src/main/java/ru/vladshi/cloudfilestorage/storage/controller/FolderController.java
@@ -12,23 +12,24 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import ru.vladshi.cloudfilestorage.security.annotation.FullPath;
 import ru.vladshi.cloudfilestorage.storage.annotation.RedirectWithPath;
 import ru.vladshi.cloudfilestorage.storage.exception.FolderAlreadyExistsException;
-import ru.vladshi.cloudfilestorage.storage.exception.StorageItemNameValidationException;
 import ru.vladshi.cloudfilestorage.storage.exception.ObjectDeletionException;
+import ru.vladshi.cloudfilestorage.storage.exception.StorageItemNameValidationException;
 import ru.vladshi.cloudfilestorage.storage.exception.StorageLimitExceededException;
 import ru.vladshi.cloudfilestorage.storage.model.FullItemPath;
-import ru.vladshi.cloudfilestorage.security.annotation.FullPath;
-import ru.vladshi.cloudfilestorage.storage.service.MinioService;
+import ru.vladshi.cloudfilestorage.storage.service.FolderService;
+import ru.vladshi.cloudfilestorage.storage.service.StorageUsageService;
 import ru.vladshi.cloudfilestorage.storage.util.DispositionHeaderUtil;
-import ru.vladshi.cloudfilestorage.storage.validation.StorageItemNameValidator;
 
 @Controller
 @RequestMapping("folder")
 @RequiredArgsConstructor
 public class FolderController {
 
-    private final MinioService minioService;
+    private final FolderService folderService;
+    private final StorageUsageService storageUsageService;
 
     @PostMapping("/create")
     @RedirectWithPath
@@ -36,8 +37,7 @@ public class FolderController {
                              @RequestParam String newFolderName,
                              RedirectAttributes redirectAttributes) throws Exception {
         try {
-            StorageItemNameValidator.validate(newFolderName);
-            minioService.createFolder(path.full(), newFolderName.strip());
+            folderService.create(path.full(), newFolderName.strip());
         } catch (FolderAlreadyExistsException | StorageItemNameValidationException e) {
             redirectAttributes.addFlashAttribute(
                     "errorMessage", "Failed to create folder. " + e.getMessage());
@@ -50,7 +50,7 @@ public class FolderController {
                              @RequestParam String folderToDelete,
                              RedirectAttributes redirectAttributes) throws Exception {
         try {
-            minioService.deleteFolder(path.full(), folderToDelete);
+            folderService.delete(path.full(), folderToDelete);
         } catch (ObjectDeletionException e) {
             redirectAttributes.addFlashAttribute(
                     "errorMessage", "Failed to delete folder. " + e.getMessage());
@@ -64,8 +64,7 @@ public class FolderController {
                              @RequestParam String newFolderName,
                              RedirectAttributes redirectAttributes) throws Exception {
         try {
-            StorageItemNameValidator.validate(newFolderName);
-            minioService.renameFolder(path.full(), folderToRename, newFolderName.strip());
+            folderService.rename(path.full(), folderToRename, newFolderName.strip());
         } catch (FolderAlreadyExistsException | ObjectDeletionException | StorageItemNameValidationException e) {
             redirectAttributes.addFlashAttribute(
                     "errorMessage", "Failed to rename folder. " + e.getMessage());
@@ -79,8 +78,8 @@ public class FolderController {
                              @RequestParam("files") MultipartFile[] files,
                              RedirectAttributes redirectAttributes) throws Exception {
         try {
-            minioService.checkStorageLimit(path.userPrefix(), files);
-            minioService.uploadFolder(path.full(), folderName, files);
+            storageUsageService.checkLimit(path.userPrefix(), files);
+            folderService.upload(path.full(), folderName, files);
         } catch (FolderAlreadyExistsException | IllegalArgumentException
                  | StorageItemNameValidationException | StorageLimitExceededException e) {
             redirectAttributes.addFlashAttribute(
@@ -92,7 +91,7 @@ public class FolderController {
     public ResponseEntity<InputStreamResource> downloadFolder(@FullPath FullItemPath path,
                                                               @RequestParam String folderName) throws Exception {
 
-        InputStreamResource folderResource = minioService.downloadFolder(path.full(), folderName);
+        InputStreamResource folderResource = folderService.download(path.full(), folderName);
 
         HttpHeaders headers = new HttpHeaders();
         headers.add(HttpHeaders.CONTENT_DISPOSITION, DispositionHeaderUtil.build(folderName) + ".zip");
