+++ src/main/java/ru/vladshi/cloudfilestorage/storage/controller/SearchController.java
@@ -6,22 +6,22 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import ru.vladshi.cloudfilestorage.storage.model.FullItemPath;
 import ru.vladshi.cloudfilestorage.security.annotation.FullPath;
-import ru.vladshi.cloudfilestorage.storage.service.MinioService;
+import ru.vladshi.cloudfilestorage.storage.model.FullItemPath;
+import ru.vladshi.cloudfilestorage.storage.service.SearchService;
 
 @Controller
 @RequestMapping("/search")
 @RequiredArgsConstructor
 public class SearchController {
 
-    private final MinioService minioService;
+    private final SearchService searchService;
 
     @GetMapping
     public String search(@FullPath FullItemPath path,
                          Model model,
                          @RequestParam(required = false) String searchQuery) throws Exception {
-        model.addAttribute("items", minioService.searchItems(path.userPrefix(), searchQuery.strip()));
+        model.addAttribute("items", searchService.searchItems(path.userPrefix(), searchQuery.strip()));
         model.addAttribute("searchQuery", searchQuery);
 
         return "search";
