+++ src/main/java/ru/vladshi/cloudfilestorage/storage/controller/StorageMainController.java
@@ -5,11 +5,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import ru.vladshi.cloudfilestorage.security.annotation.FullPath;
 import ru.vladshi.cloudfilestorage.storage.model.FullItemPath;
 import ru.vladshi.cloudfilestorage.storage.model.StorageItem;
-import ru.vladshi.cloudfilestorage.storage.model.UserStorageInfo;
-import ru.vladshi.cloudfilestorage.security.annotation.FullPath;
-import ru.vladshi.cloudfilestorage.storage.service.MinioService;
+import ru.vladshi.cloudfilestorage.storage.model.StorageUsageInfo;
+import ru.vladshi.cloudfilestorage.storage.service.FolderService;
+import ru.vladshi.cloudfilestorage.storage.service.StorageUsageService;
 import ru.vladshi.cloudfilestorage.storage.util.BreadcrumbUtil;
 
 import java.util.List;
@@ -19,14 +20,14 @@ import java.util.List;
 @RequiredArgsConstructor
 public class StorageMainController {
 
-    private final MinioService minioService;
+    private final FolderService folderService;
+    private final StorageUsageService storageUsageService;
 
     @GetMapping
-    public String showFiles(@FullPath FullItemPath path,
-                            Model model) throws Exception {
+    public String showFiles(@FullPath FullItemPath path, Model model) throws Exception {
 
-        List<StorageItem> storageItems = minioService.getItems(path.full());
-        UserStorageInfo storageInfo = minioService.getUserStorageInfo(path.userPrefix());
+        List<StorageItem> storageItems = folderService.getFolderContents(path.full());
+        StorageUsageInfo storageInfo = storageUsageService.getInfo(path.userPrefix());
 
         model.addAttribute("path", path.relativePath());
         model.addAttribute("breadcrumbs", BreadcrumbUtil.buildBreadcrumbs(path.relativePath()));
