+++ src/main/java/ru/vladshi/cloudfilestorage/storage/service/SearchService.java
@@ -1,0 +1,11 @@
+package ru.vladshi.cloudfilestorage.storage.service;
+
+import ru.vladshi.cloudfilestorage.storage.model.StorageItem;
+
+import java.util.List;
+
+public interface SearchService {
+
+    List<StorageItem> searchItems(String userPrefix, String query) throws Exception;
+
+}
