+++ src/main/java/ru/vladshi/cloudfilestorage/user/service/impl/UserServiceImpl.java
@@ -6,11 +6,11 @@ import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import ru.vladshi.cloudfilestorage.storage.service.FolderService;
+import ru.vladshi.cloudfilestorage.storage.service.UserPrefixService;
 import ru.vladshi.cloudfilestorage.user.entity.User;
 import ru.vladshi.cloudfilestorage.user.exception.UserRegistrationException;
 import ru.vladshi.cloudfilestorage.user.repository.UserRepository;
-import ru.vladshi.cloudfilestorage.storage.service.MinioService;
-import ru.vladshi.cloudfilestorage.storage.service.UserPrefixService;
 import ru.vladshi.cloudfilestorage.user.service.UserService;
 
 @Service
@@ -21,7 +21,7 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
     private final UserPrefixService userPrefixService;
-    private final MinioService minioService;
+    private final FolderService folderService;
 
     @Override
     public void registerUser(User user) {
@@ -29,7 +29,7 @@ public class UserServiceImpl implements UserService {
         try {
             userRepository.save(user);
             String userFolderName = userPrefixService.buildUserPrefix(user.getId(), user.getUsername());
-            minioService.createUserRootFolder(userFolderName);
+            folderService.createUserRootFolder(userFolderName);
         } catch (DataIntegrityViolationException e) {
             if (e.getCause() instanceof ConstraintViolationException) {
                 throw new UserRegistrationException("Username is already taken");
